refactor: rename redux-states/reducer.ts to root-reducer.ts
