+++ src/redux-states/store.ts
@@ -1,7 +1,7 @@
 import { applyMiddleware, createStore } from 'redux';
 import thunk from 'redux-thunk';
 
-import rootReducer from './reducer';
+import rootReducer from './root-reducer';
 
 const store = createStore(rootReducer, applyMiddleware(thunk));
